refactor: update import paths and rename core-related files for consistency
