+++ apps/electron/src/main/infrastructure/database/base-repository.ts
