+++ apps/electron/src/main/infrastructure/database/sqlite-manager.ts
