+++ apps/electron/src/main/modules/agent/agent-repository.ts
@@ -1,8 +1,8 @@
-import { BaseRepository } from "../../infrastructure/database/core/base-repository";
+import { BaseRepository } from "../../infrastructure/database/base-repository";
 import {
   SqliteManager,
   getSqliteManager,
-} from "../../infrastructure/database/core/sqlite-manager";
+} from "../../infrastructure/database/sqlite-manager";
 import { AgentConfig } from "@mcp_router/shared";
 import { v4 as uuidv4 } from "uuid";
