+++ apps/electron/src/main/modules/agent/deployed-agent-repository.ts
@@ -1,9 +1,9 @@
 import { DeployedAgent } from "@mcp_router/shared";
-import { BaseRepository } from "../../infrastructure/database/core/base-repository";
+import { BaseRepository } from "../../infrastructure/database/base-repository";
 import {
   SqliteManager,
   getSqliteManager,
-} from "../../infrastructure/database/core/sqlite-manager";
+} from "../../infrastructure/database/sqlite-manager";
 import { v4 as uuidv4 } from "uuid";
 
 /**
