+++ apps/electron/src/main/modules/mcp-apps-manager/token-repository.ts
@@ -1,8 +1,8 @@
-import { BaseRepository } from "../../infrastructure/database/core/base-repository";
+import { BaseRepository } from "../../infrastructure/database/base-repository";
 import {
   SqliteManager,
   getSqliteManager,
-} from "../../infrastructure/database/core/sqlite-manager";
+} from "../../infrastructure/database/sqlite-manager";
 import { Token } from "@mcp_router/shared";
 
 /**
