+++ apps/electron/src/main/modules/workspace/database-context.ts
@@ -1,4 +1,4 @@
-import { SqliteManager } from "../../infrastructure/database/core/sqlite-manager";
+import { SqliteManager } from "../../infrastructure/database/sqlite-manager";
 
 /**
  * Database context manager class
