+++ apps/electron/src/main/modules/workspace/workspace.repository.ts
@@ -1,8 +1,8 @@
-import { BaseRepository } from "../../infrastructure/database/core/base-repository";
+import { BaseRepository } from "../../infrastructure/database/base-repository";
 import {
   SqliteManager,
   getSqliteManager,
-} from "../../infrastructure/database/core/sqlite-manager";
+} from "../../infrastructure/database/sqlite-manager";
 import { Workspace } from "@mcp_router/shared";
 
 export class WorkspaceRepository extends BaseRepository<Workspace> {
